+++ nextcord/voice_recording/opus.py
@@ -1,0 +1,70 @@
+# SPDX-License-Identifier: MIT
+
+from __future__ import annotations
+
+import gc
+import logging
+from threading import Thread, Event
+from time import sleep
+from typing import TYPE_CHECKING, Dict
+
+from nextcord import opus
+from nextcord.errors import InvalidArgument
+
+if TYPE_CHECKING:
+    from .core import (
+        OpusFrame,
+        RecorderClient
+    )
+
+class DecoderThread(Thread, opus._OpusStruct):
+    def __init__(self, recorder) -> None:
+        super().__init__(daemon=True)
+
+        self.recorder: RecorderClient = recorder
+        self.decode_queue: list[OpusFrame] = []
+        self.decoder: Dict[int, opus.Decoder] = {}
+        self._end = Event()
+
+    def start(self):
+        self._end = Event()
+        super().start()
+
+    def decode(self, opus_frame) -> None:
+        self.decode_queue.append(opus_frame)
+
+    def run(self) -> None:
+        while not self._end.is_set():
+            try:
+                opus_frame: OpusFrame = self.decode_queue.pop(0)
+                
+            except IndexError:
+                sleep(0.001)
+                continue
+
+            try:
+                if opus_frame.decrypted_data is None:
+                    continue
+                else:
+                    decoder = self.get_decoder(opus_frame.ssrc)
+                    opus_frame.decoded_data = decoder.decode(
+                        opus_frame.decrypted_data, fec=False
+                    )
+            except InvalidArgument:
+                print("Error occurred while decoding opus frame.")
+                continue
+
+            self.recorder._process_decoded_audio(opus_frame)
+
+    def stop(self) -> None:
+        self._end.set()
+        self.decode_queue.clear()
+        self.decoder.clear()
+        gc.collect()
+        logging.debug("Decoder process killed.")
+
+    def get_decoder(self, ssrc) -> opus.Decoder:
+        if (d := self.decoder.get(ssrc)) is not None:
+            return d
+        d = self.decoder[ssrc] = opus.Decoder()
+        return d
