--- nextcord/voicerec/__init__.py
@@ -1,3 +1,0 @@
-from .core import *
-from .errors import *
-from .exporters import *
