--- nextcord/voicerec/opus.py
@@ -1,61 +1,0 @@
-# SPDX-License-Identifier: MIT
-
-
-import gc
-import logging
-import threading
-from time import sleep
-
-from nextcord import opus
-from nextcord.errors import InvalidArgument
-
-
-class DecoderThread(threading.Thread, opus._OpusStruct):
-    def __init__(self, recorder) -> None:
-        super().__init__(daemon=True)
-
-        self.recorder = recorder
-        self.decode_queue = []
-        self.decoder = {}
-        self._end_thread = threading.Event()
-
-    def start(self) -> None:
-        self._end_thread = threading.Event()
-        super().start()
-
-    def decode(self, opus_frame) -> None:
-        self.decode_queue.append(opus_frame)
-
-    def run(self) -> None:
-        while not self._end_thread.is_set():
-            try:
-                opus_frame = self.decode_queue.pop(0)
-
-            except IndexError:
-                sleep(0.001)
-                continue
-
-            try:
-                if opus_frame.decrypted_data is None:
-                    continue
-                else:
-                    decoder = self.get_decoder(opus_frame.ssrc)
-                    opus_frame.decoded_data = decoder.decode(opus_frame.decrypted_data, fec=False)
-            except InvalidArgument:
-                print("Error occurred while decoding opus frame.")
-                continue
-
-            self.recorder._process_decoded_audio(opus_frame)
-
-    def stop(self) -> None:
-        self._end_thread.set()
-        self.decode_queue.clear()
-        self.decoder.clear()
-        gc.collect()
-        logging.debug("Decoder process killed.")
-
-    def get_decoder(self, ssrc) -> opus.Decoder:
-        if (d := self.decoder.get(ssrc)) is not None:
-            return d
-        d = self.decoder[ssrc] = opus.Decoder()
-        return d
